+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fixed Loremaster to work with Isekai Protagonist and Villain Spellbook.
+
 ## [1.8.0] - 2022-11-13
 ### Added
 - Added `Isekai Protagonist` dialogue when speaking to Hulrun at the Kenabres festival.